With @travs: Continuing breakdown of test.py into individuals tests.
