+++ tests/NeuroMLTest.py
@@ -1,0 +1,26 @@
+import sys
+sys.path.insert(0,".")
+import unittest
+import neuroml
+import neuroml.writers as writers
+import PyOpenWorm
+from PyOpenWorm import *
+import networkx
+import rdflib
+import rdflib as R
+import pint as Q
+import os
+import subprocess as SP
+import subprocess
+import tempfile
+import doctest
+
+from glob import glob
+
+from GraphDBInit import * 
+
+from DataTestTemplate import _DataTest
+
+class NeuroMLTest(_DataTest):
+    """What is this test supposed to do?"""
+    pass
